LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-range-ignored-expected.txt


FAIL Animation API call rangeStart overrides animation-range-start assert_approx_equals: values do not match for "timeline's current time before style change" expected 16.666666666666668 +/- 0.125 but got 13.333333969116211
FAIL Animation API call rangeEnd overrides animation-range-end assert_approx_equals: values do not match for "timeline's current time after first set of range updates" expected 16.666666666666668 +/- 0.125 but got 13.333333969116211
